+++ src/screens/Details/index.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import {View, Text} from 'react-native'
+
+export default function Details(){
+    return(
+        <View>
+            <Text>Essa é a tela de detalhes</Text>
+        </View>
+    )
+}
